+++ R/archeoViz.R
@@ -20,8 +20,8 @@ archeoViz <- function(objects.df=NULL, refits.df=NULL, timeline.df=NULL,
   }
   
   # : test lang ----
-  if( ! lang %in% c("en", "fr", "pt")){
-    stop("The 'lang' parameter must be one of 'en', 'fr', 'pt'.")
+  if( ! lang %in% c("en", "fr", "it", "pt")){
+    stop("The 'lang' parameter must be one of 'en', 'fr', 'it, 'pt'.")
   }
   
   # : test square.size ----